% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-deploy.R
\name{build_lambda}
\alias{build_lambda}
\title{build and tag lambda image locally}
\usage{
build_lambda(tag, runtime_function, runtime_path, dependencies)
}
\arguments{
\item{tag}{A name for the Docker container and Lambda function}

\item{runtime_function}{name of the runtime function}

\item{runtime_path}{path to the script containing the runtime function}

\item{dependencies}{list of dependencies}
}
\description{
build and tag lambda image locally
}
