% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-util.R
\name{create_lambda_dockerfile}
\alias{create_lambda_dockerfile}
\title{create_lambda_dockerfile}
\usage{
create_lambda_dockerfile(folder, runtime_function, runtime_path, dependencies)
}
\arguments{
\item{folder}{path to store the Dockerfile}

\item{runtime_function}{name of the runtime function}

\item{runtime_path}{path to the script containing the runtime function}

\item{dependencies}{list of dependencies}
}
\description{
create_lambda_dockerfile
}
\examples{
\dontrun{
  runtime_function <- "parity"
  runtime_path <- system.file("parity.R", package = "r2lambda")
  folder <- "~/Desktop/parity-lambda"
  dependencies <- NULL

  create_lambda_dockerfile(
    folder = folder,
    runtime_function = runtime_function,
    runtime_path = runtime_path,
    dependencies = dependencies
    )
  dir.exists(folder)
  dir(folder)
}
}
