% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-util.R
\name{create_lambda_exec_role}
\alias{create_lambda_exec_role}
\title{create_lambda_exec_role}
\usage{
create_lambda_exec_role(tag)
}
\arguments{
\item{tag}{the tag of an existing local image}
}
\description{
create_lambda_exec_role
}
