% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-util.R
\name{create_lambda_function}
\alias{create_lambda_function}
\title{create_lambda_function}
\usage{
create_lambda_function(tag, ecr_image_uri, lambda_role_arn)
}
\arguments{
\item{tag}{the tag of an existing local image}

\item{ecr_image_uri}{the URI of the image to use}

\item{lambda_role_arn}{the arn of the execution role created for this lambda}
}
\description{
create_lambda_function
}
