% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-util.R
\name{create_lambda_image}
\alias{create_lambda_image}
\title{create_lambda_container}
\usage{
create_lambda_image(folder, tag)
}
\arguments{
\item{folder}{path to the folder containing the lambda runtime script and Dockerfile}

\item{tag}{a tag for the image}
}
\description{
create_lambda_container
}
