% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-deploy.R
\name{invoke_lambda}
\alias{invoke_lambda}
\title{invoke a lambda function}
\usage{
invoke_lambda(function_name, invocation_type, payload, include_logs = FALSE)
}
\arguments{
\item{function_name}{The name or arn of the function}

\item{invocation_type}{One of ‘DryRun’, ‘RequestResponse’, or ‘Event’ see ?paws.compute::lambda_invoke}

\item{payload}{A named list internally converted to json}

\item{include_logs}{logical, whether to show the lambda logs (default: FALSE)}
}
\description{
invoke a lambda function
}
\examples{
\dontrun{
  # assumes a lambda called parity is deployed
  invoke_lambda(
   function_name = "lambda-test88",
   payload = list(number = 3),
   invocation_type = "RequestResponse"
  )
}
}
