% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-util.R
\name{push_lambda_image}
\alias{push_lambda_image}
\title{push_lambda_container to AWS ECR}
\usage{
push_lambda_image(tag)
}
\arguments{
\item{tag}{the tag of an existing local image}
}
\description{
push_lambda_container to AWS ECR
}
