% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-deploy.R
\name{test_lambda}
\alias{test_lambda}
\title{test a lambda locally}
\usage{
test_lambda(tag, payload)
}
\arguments{
\item{tag}{The tag of an existing local image tagged with ECR repo (see `build_lambda`)}

\item{payload}{Named list. Arguments to lambda function.}
}
\description{
test a lambda locally
}
\examples{

payload <- list(number = 2)
tag <- "449283523352.dkr.ecr.us-east-1.amazonaws.com/myrepo51:latest"
test_lambda(tag, payload)

}
