% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-util.R
\name{aws_connect}
\alias{aws_connect}
\title{connect to an aws service}
\usage{
aws_connect(service)
}
\arguments{
\item{service}{character, the name of a service, e.g., "lambda" or "iam". Should
be a function exported by `{paws}` (see `getNamespaceExports("paws")`)}
}
\description{
connect to an aws service
}
\examples{
\dontrun{
  aws_connect("lambda")
  }
}
