% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-deploy.R
\name{deploy_lambda}
\alias{deploy_lambda}
\title{deploy a local lambda image to AWS Lambda}
\usage{
deploy_lambda(tag)
}
\arguments{
\item{tag}{The tag of an existing local image tagged with ECR repo (see `build_lambda`)}
}
\description{
deploy a local lambda image to AWS Lambda
}
\examples{
\dontrun{

  runtime_function <- "parity"
  runtime_path <- system.file("parity.R", package = "r2lambda")
  dependencies <- NULL

  build_lambda(
    tag = "myrepo45",
    runtime_function = runtime_function,
    runtime_path = runtime_path,
    dependencies = dependencies
    )

  deploy_lambda(tag = "myrepo45")

  invoke_lambda(
    function_name = "myrepo45",
    payload = list(number = 3),
    invocation_type = "RequestResponse"
    )

}
}
