% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda-deploy.R
\name{deploy_lambda}
\alias{deploy_lambda}
\title{deploy an R script to AWS Lambda}
\usage{
deploy_lambda(tag, runtime_function, runtime_path, dependencies)
}
\arguments{
\item{tag}{A name for the Docker container and Lambda function}

\item{runtime_function}{name of the runtime function}

\item{runtime_path}{path to the script containing the runtime function}

\item{dependencies}{list of dependencies}
}
\description{
deploy an R script to AWS Lambda
}
\examples{
\dontrun{

  runtime_function <- "parity"
  runtime_path <- system.file("parity.R", package = "r2lambda")
  dependencies <- NULL

  deploy_lambda(
    tag = "myrepo45",
    runtime_function = runtime_function,
    runtime_path = runtime_path,
    dependencies = dependencies
    )

 invoke_lambda(
   function_name = "myrepo45",
   payload = list(number = 3),
   invocation_type = "RequestResponse"
  )

}
}
